e2e - move func to `helper.go`
